requirements.txt
gradio 
PyMuPDF 
python-docx 
pandas 
openpyxl
sentence-transformers 
faiss-cpu
langchain 
sqlalchemy 
pymysql
langchain 
transformers